backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.4.0
scipy==1.11.4
joblib==1.3.2
redis==5.0.1
celery==5.3.4
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.2
auto-sklearn==0.15.0
